[FIX] #59450 Invoice Time Prices

+++ dhHelpdeskNG/dhHelpdesk_NG.Web/Areas/Invoices/Models/LogInvoiceItemViewModel.cs
@@ -13,8 +13,8 @@ namespace DH.Helpdesk.Web.Areas.Invoices.Models
 		public string Text { get; set; }
 		public int WorkingTime { get; set; }
 		public int Overtime { get; set; }
-		public int Material { get; set; }
-		public decimal Price { get; set; }
+		public decimal Material { get; set; }
+		public int Price { get; set; }
 		public bool Charge { get; set; }
 
 		public int WorkingHourRate { get; set; }
